+++ packages/cloud-cognitive/src/components/CreateSidePanel/CreateSidePanel.stories.js
@@ -21,7 +21,6 @@ import {
 } from 'carbon-components-react/lib/components/UIShell';
 
 import { pkg } from '../../settings';
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { getStorybookPrefix } from '../../../config';
 
 import { CreateSidePanel } from '.';