+++ packages/cloud-cognitive/src/components/HTTPErrors/HTTPError403/HTTPError403.stories.js
@@ -7,7 +7,6 @@
 
 import React from 'react';
 import { pkg } from '../../../settings';
-import '../../../utils/enable-all'; // must come before component is imported
 import { HTTPError403 } from '.';
 import { getStorybookPrefix } from '../../../../config';
 const storybookPrefix = getStorybookPrefix(pkg, HTTPError403.displayName);