+++ packages/cloud-cognitive/src/components/Saving/Saving.stories.js
@@ -8,7 +8,6 @@
 import React, { useState, useEffect } from 'react';
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 import { pkg } from '../../settings';
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { getStorybookPrefix } from '../../../config';
 import { Saving } from '.';
 import mdx from './Saving.mdx';