+++ packages/cloud-cognitive/src/components/TagSet/TagSet.test.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { pkg, carbon } from '../../settings';
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { TagSet } from '.';
 import { TagSetModal } from './TagSetModal';
 import { Tag } from 'carbon-components-react';